requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
cssutils==2.4.0
Django==3.2
django-bootstrap4==22.1
django-crispy-forms==1.14.0
django-filter==21.1
django-inlinecss==0.3.0
future==0.18.2
pynliner==0.8.0
PySocks==1.7.1
pytz==2022.1
soupsieve==2.3.1
sqlparse==0.4.2
xlwt==1.3.0
